rxgen: Use a function to create the rxInterface
rx_parse uses a string template that references the serviceID and the
servicePort of a call connection when building a call to
rx_RecordCallStatistics.

Replace the direct member references in the template with a call to a
new function, rx_GetCallInterface, that returns the same calculated
value.  This will isolate any future changes to the template string that
might be needed for the IPV6 support.

Change-Id: Ifd31962822522157a3b0a8eeb9081f3cd4fb8061