butc: fix int to float conversion warning
Building with clang-10 results in 2 warnings/errors associated with
with trying to convert 0x7fffffff to a floating point value.

    tcmain.c:240:18: error: implicit conversion from 'int' to 'float'
                 changes value from 2147483647 to 2147483648 [-Werror,
                 -Wimplicit-int-float-conversion]
    if ((total > 0x7fffffff) || (total < 0))    /* Don't go over 2G */

Increase the constant by 1, which converts cleanly from an int to a
float and change the test from '>' to '>='.

Assign the max value directly to the return variable to avoid the second
conversion.