Provide numerical value for AFS_MOUNT_AFS
The define AFS_MOUNT_AFS is either a numerical constant or the string
"afs".  The string cannot be used in expressions that are expecting
a numerical value.

Define a new macro AFS_MOUNT_AFS_N which is set to the value 0x616673
('afs') so it can be used in expressions.

Change-Id: I909570b48329350dd35b6dcd9f1cf114d866d2de