afs: f_fsid set to incorrect value
Building with clang-10 produces the warning/error message
    warning: converting the result of '<<' to a boolean always evaluates
    to true [-Wtautological-constant-compare]
for the expression
    abp->f_fsid = (AFS_VFSMAGIC << 16) || AFS_VFSFSID;

The message is because a logical OR '||' is used instead of a bitwise
OR '|'.  The result of this expression will always set the f_fsid
member to a 1 and not the indented value of AFS_VFSMAGIC combined with
AFS_VFSFSID.

Update the expression to use a bitwise OR instead of the logical OR.

There is a complication in that the macro AFS_VFSFSID can be a numeric
constant, a variable, or a string depending on which
'src/config/param...h' is used and the platform environment.  If the
macro AFS_MOUNT_AFS_N is defined, set a new macro AFS_VFSFSID_V with
the expression to be used to assign to f_fsid.

Clean up the preprocessor indentation in src/afs/afs/h in the area
around the AFS_VFSFSID defines.

NOTE. This will change value stored in the f_fsid that is returned from
statfs.

Change-Id: I618fc4c89029a6cca2ca6f530b8f65399299a9d1