clang-10: use AFS_FALLTHROUGH for case fallthrough
Clang-10 will not recognize '/* fallthrough */' as an indicator to
turn off the fallthrough warning due to the lack of a 'break' in a case
block.

In addition clang-10 is finding additional locations where fall throughs
occur.

Replace the '/* fallthrough */' comments with AFS_FALLTHROUGH.  Add
AFS_FALLTHROUGH as needed.

Correct a mis-indented brace '}' in src/WINNT/afsd/smb3.c

Note, the clang maintainers have rejected the use of comments as a flag
to turn off the fall through warnings.

Change-Id: I8b85aa2eba9cfd222ec7e995c1f9e070fde612cb