clang-10: ignore fallthrough warning in generated code
Clang-10 will not recognize '/* fall through */' as an indicator to
turn off the fallthrough warning due to the lack of a 'break' in a case
block.

Code generated by flex uses the '/* fall through */' comments to turn
off compiler warnings for fall throughs in case blocks.

Use the flex %top{..} section to insert a #pragma to ignore the
-Wimplicit-fallthrough diagnostics if clang is being used.

Change-Id: I4c054defda03daa2aeb645ae2271dfa0cb54925f