autoconf: fix detection for fallthrough attribute
The external source 'autoconf-archive/m4/ax_gcc_func_attribute.m4' fails
to properly detect the fallthrough attribute due to other warnings
flagging the switch statement's a value that is not in a case and a
mismatch on the return with the function's signature.

Update ax_gcc_func_attribute.m4 to fix the warnings that are masking
the detection of the fallthrough attribute.

Note that the problem has been reported with a patch to the
autoconf-archive maintainers.

Change-Id: I80a4557384f8e1438344e48bfe722e20c8773882