volser: Always fetch locked entry in CheckVolume
CheckVolume accepts an optional nvldbentry as an argument.  When this
argument is NULL, CheckVolume correctly locks the nvldbentry before
obtaining the entry and making changes.  However, if an nvldbentry is
passed, CheckVolume merely copies it to a work entry; it also neglects
to obtain the current state of the entry after locking it for
modification.

Change CheckVolume to obtain a locked copy of the nvldbentry if
modifications are required.

Reviewed-on: https://gerrit.openafs.org/14356
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
(cherry picked from commit c20dceb39292903595bb2f0f995ccf6754e3841c)

Change-Id: I48f022536224b94dfa9b986cff33075bfc4a230b