ptserver: don't allow new blank users or groups
Currently, the ptserver does not prevent the creation of blank users or
groups. Since users / groups without a name are not desirable, check if
the name provided by the user is not empty. Notice that this check is
exclusive for new users / groups so blank entries created in the past
can still be manipulated (e.g. removed).

Change-Id: Iea321d67e73b6a9bd409c573bca54f10b8cadf34