rxkad: Fix NULL Dereference warnings
The static anlysis infer utility reports 'Null Dereference' errors in
the function 'der_copy_octet_string' which is implemented in
src/crypto/rfc3961/copy.c and duplicated in src/rxkad/v5der.c.

The existing logic assumes that malloc returns a NULL if the length of
the requested data is 0.  A check is made for a length > 0 and a NULL
from malloc which results in returning an error.

However if the length requested is zero, code following the check
continues with references to the pointer (but the usage is guarded by
testing the length).  While logically this is correct it does assume
that all references are guarded by testing the length first.  It also
causes some static analysis utilities to report it as a NULL
dereference.

There are additional functions with occurrences of this usage pattern in
v5der.c that the infer analysis is not reporting.

Update copy.c and v5der.c to add an explicit check for a NULL pointer
and skip the code paths that are guarded by relying on a length of 0.

This update should be functionally equivalent.

Note: there is an existing assumption that the caller of these functions
are correctly handling the case where the length is 0 and the pointer
is NULL.

Change-Id: Ib65718f5f3ac3f36302321267b045be8820e233f