libacl 13089 libacl: Document freeing acls
The method that memory for acls are allocated and freed confuses some
static analysis utilities.  Add a comment to simply document where the
memory is being released.

(cppcheck / scan-build)

Change-Id: I5392bd31a18a52273ab6e49ed3558fda2f78690e

libacl 13196 Add checks to ensure *alloc was successful

A number of locates were missing a followup check to ensure the memory
allocation was successful before attempting to utilize the memory.
(via facebook-infer)

Change-Id: I5747a85f76d4769869858b6f8c600a2e192ced11