afsweb: Fix problems found by static analysis
Several static analysis tools have identified various problems:
 - files left open from early exits               (infer)
 - possible array overrun                         (cppcheck, infer)

To resolve the above problems:
 - close files before exiting
 - ensure array access stays in bounds

When t == 80, this function attempts to access W[80] which was out
of bounds.  W[80] is now defined with a default value of 0.

This commit is a reorganization of commits developed by Pat Riehecky,
who ran the static analysis tools and developed the fixes.

Change-Id: I5dbfc628a5dbdd7e6672ef213eca733fd3e77f79