FBSD: Use vrefl() when available
Commit 81ea6544 (FBSD: avoid vrefl()) removed our reference to
vrefl(), since it was only introduced with FreeBSD 11.0. However, it
was replaced with calls to vref() and vrele(), the latter of which can
lock the vnode, and generally is allowed to sleep.

Many osi_vnhold callers hold AFS_GLOCK, which is a non-sleepable lock,
so this can cause a panic if the vnode is VI_DOOMED and locked by
another thread.

To avoid this on at least modern FreeBSD releases, use vrefl() when it
is available (since <https://reviews.freebsd.org/D4953>).

Reviewed-on: https://gerrit.openafs.org/14796
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
(cherry picked from commit 8495165c3d5ff3167e1e3e77ade95af4be6b3895)

Change-Id: Ic54d5360aeb79fe7df1dbe165d9de10f70b3052e