AIX: Avoid COMPAT_43 for clang
We currently define COMPAT_43 for AIX. This causes some socket-related
structures and functions to be defined to be compatible with BSD 4.3,
but causes us to not get function prototypes for recvmsg() and
sendmsg(). When using the clang-based XLC 17.1, this causes errors:

    .../src/rx/rx_lwp.c:414:12: error: call to undeclared function 'recvmsg'; ISO C99 and later do not support implicit function declarations [-Wimplicit-function-declaration]
        code = recvmsg(socket, msg_p, flags);
               ^
    .../src/rx/rx_lwp.c:432:12: error: call to undeclared function 'sendmsg'; ISO C99 and later do not support implicit function declarations [-Wimplicit-function-declaration]
        while (sendmsg(socket, msg_p, flags) == -1) {
               ^

To avoid this, don't define COMPAT_43. To try to make sure we don't
change anything when using the old xlc, still define COMPAT_43 when
we're not using clang.

Reviewed-on: https://gerrit.openafs.org/15445
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
(cherry picked from commit 8207d4d74e6b243ace853f2c5fb6d44f22c7710f)

Change-Id: Ifb1ddb117456c7674a07c2ebd692ac3e88428529