viced: Protect vlConn in TryLocalVLServer
In TryLocalVLServer(), 'vlConn' is a static rx_connection that is
initialized just once. But no locks are held in this function, and so
theoretically one thread could try to use vlConn while another is
setting it, which is not safe.

This function is only used for RXAFS_GetVolumeInfo, an old RPC that
isn't called by anything in OpenAFS. Maybe this RPC can be disabled
entirely, but for now at least protect vlConn with FS_LOCK to make
sure it doesn't cause a segfault or memory corruption.

Some other races are possible with 'down' and 'lastDownTime', but
these just determine whether we exit early, and so racing on those is
not such a big problem.

FIXES 135444

Reviewed-on: https://gerrit.openafs.org/16146
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
(cherry picked from commit 035cb883454eb5a6fac3dc1e3f8ef17a9f3f169e)

Change-Id: I1210f01cc42ae563c0b62aa9788018015dadc06c